+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/MainActivity.kt
@@ -43,7 +43,9 @@ import com.mapbox.maps.extension.style.sources.generated.geoJsonSource
 import com.mapbox.maps.extension.style.sources.getSourceAs
 import com.mapbox.maps.extension.style.style
 import com.mapbox.search.MapboxSearchSdk
+import com.mapbox.search.OfflineSearchEngineSettings
 import com.mapbox.search.ResponseInfo
+import com.mapbox.search.SearchEngineSettings
 import com.mapbox.search.record.HistoryRecord
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.result.SearchSuggestion
@@ -125,7 +127,14 @@ class MainActivity : AppCompatActivity() {
         }
 
         searchResultsView = findViewById<SearchResultsView>(R.id.search_results_view).apply {
-            initialize(CommonSearchViewConfiguration(DistanceUnitType.IMPERIAL))
+            val context = this@MainActivity
+            initialize(
+                SearchResultsView.Configuration(
+                    commonConfiguration = CommonSearchViewConfiguration(DistanceUnitType.IMPERIAL),
+                    searchEngineSettings = SearchEngineSettings(context, BuildConfig.MAPBOX_API_TOKEN),
+                    offlineSearchEngineSettings = OfflineSearchEngineSettings(context, BuildConfig.MAPBOX_API_TOKEN)
+                )
+            )
             isVisible = false
         }