+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/ReverseGeocodingJavaExampleActivity.java
@@ -13,8 +13,10 @@ import com.mapbox.search.ResponseInfo;
 import com.mapbox.search.ReverseGeoOptions;
 import com.mapbox.search.SearchCallback;
 import com.mapbox.search.SearchEngine;
+import com.mapbox.search.SearchEngineSettings;
 import com.mapbox.search.SearchRequestTask;
 import com.mapbox.search.result.SearchResult;
+import com.mapbox.search.sample.BuildConfig;
 
 import java.util.List;
 
@@ -44,7 +46,9 @@ public class ReverseGeocodingJavaExampleActivity extends AppCompatActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        searchEngine = MapboxSearchSdk.getSearchEngine();
+        searchEngine = MapboxSearchSdk.createSearchEngineWithBuiltInDataProviders(
+            new SearchEngineSettings(this, BuildConfig.MAPBOX_API_TOKEN)
+        );
 
         final ReverseGeoOptions options = new ReverseGeoOptions.Builder(Point.fromLngLat(2.294434, 48.858349))
             .limit(1)